pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "vlarl"
authors = [
    {name = "Guanxing Lu"},
    {name = "Chubin Zhang"},
    {name = "Haonan Jiang"},
    {name = "Yuheng Zhou"},
    {name = "Zifeng Gao"},
    {name = "Yansong Tang"},
    {name = "Ziwei Wang"},
]
description = "VLA-RL: Towards Masterful and General Robotic Manipulation with Scalable Reinforcement Learning"
version = "0.0.3"
readme = "README.md"
requires-python = ">=3.10,<3.11"
keywords = ["vision-language-action", "reinforcement learning", "robotic manipulation"]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "accelerate",
    "draccus",
    "einops",
    "huggingface_hub",
    "json-numpy",
    "jsonlines",
    "matplotlib",
    "peft",
    "protobuf",
    "rich",
    "sentencepiece",
    "timm==0.9.10",
    "tokenizers",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "transformers==4.48.2",
    "wandb",
    "tensorflow",
    "tensorflow_datasets",
    "tensorflow_graphics",
    "dlimp @ git+https://github.com/moojink/dlimp_openvla.git",
    "pyzmq",
    "vllm==0.7.2",
    "ray==2.40.0",
    "cupy-cuda12x",
    "pandas",
    "termcolor",
    "tensorboard",
    "gymnasium==0.29.1",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
    "gpustat",
    "ipython",
    "pre-commit",
    "ruff>=0.2.2",
]
sagemaker = [
    "boto3",
    "sagemaker"
]

[project.urls]
homepage = "https://github.com/GuanxingLu/vlarl"
repository = "https://github.com/GuanxingLu/vlarl"
documentation = "https://github.com/GuanxingLu/vlarl"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["cache"]

[tool.setuptools.package-data]
"prismatic" = ["py.typed"]

[tool.black]
line-length = 121
target-version = ["py38", "py39", "py310"]
preview = true

[tool.ruff]
line-length = 121
target-version = "py38"

[tool.ruff.lint]
select = ["A", "B", "E", "F", "I", "RUF", "W"]
ignore = ["F722"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.uv]
dev-dependencies = [
    "black>=24.2.0",
    "gpustat",
    "ipython",
    "pre-commit",
    "ruff>=0.2.2",
]
